requirements.txt
accelerate==0.21.0
bitsandbytes==0.43.1
gradio==4.33.0
iopath==0.1.10
matplotlib==3.8.4
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
opencv_contrib_python==4.5.4.60
opencv_python==4.5.4.60
opencv_python_headless==4.5.4.60
pandas==2.2.2
peft==0.10.0
Pillow==10.3.0
PyYAML==6.0.1
Requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
scikit-image==0.23.2
timm==0.6.13
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.2
transformers==4.41.2
visual_genome==1.1.1
wandb==0.16.6
webdataset==0.2.86
sentencepiece==0.2.0